feat: add BigDecimal

+++ src/banco/modelo/pagamento/DocumentoPagavel.java
@@ -1,7 +1,9 @@
 package banco.modelo.pagamento;
 
+import java.math.BigDecimal;
+
 public interface DocumentoPagavel {
-    double getValorTotal();
+    BigDecimal getValorTotal();
     boolean estaPago();
     void quitarPagamento();
     default void imprimirRecibo() {
